Refactor project folder structure; moved all sources to src/ but main.py.
