+++ README.md
@@ -1,0 +1,5 @@
+# Assignment for **Análisis Masivo de Datos** subject's part 2; data graph
+
+Marcos Barrios, Eugenio Hernandez
+
+
